data/requirements.txt
Django
django-piston